Move BBB to certificates app
