+++ website/borrel/models.py
@@ -15,19 +15,6 @@ from venues.models import Reservation
 User = get_user_model()
 
 
-class BasicBorrelBrevet(models.Model):
-    """Basic Borrel Brevet class."""
-
-    user = models.OneToOneField(
-        User, blank=False, null=False, on_delete=models.CASCADE, related_name="basic_borrel_brevet"
-    )
-    registered_on = models.DateField(auto_now_add=True, blank=False, null=False)
-
-    def __str__(self):
-        """Convert this object to string."""
-        return f"Borrel brevet for {self.user}"
-
-
 class ProductCategory(models.Model):
     """Product category model."""
